+++ x-pack/plugins/spaces/public/copy_saved_objects_to_space/components/index.ts
@@ -4,5 +4,3 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-
-export { CopySavedObjectsToSpaceFlyout } from './copy_to_space_flyout';
