+++ x-pack/plugins/spaces/public/management/spaces_management_app.tsx
@@ -5,18 +5,11 @@
  * 2.0.
  */
 
-import React from 'react';
-import { render, unmountComponentAtNode } from 'react-dom';
-import { Router, Route, Switch, useParams } from 'react-router-dom';
 import { i18n } from '@kbn/i18n';
 import { StartServicesAccessor } from 'src/core/public';
-import { RedirectAppLinks } from '../../../../../src/plugins/kibana_react/public';
 import { RegisterManagementAppArgs } from '../../../../../src/plugins/management/public';
 import { PluginsStart } from '../plugin';
-import { SpacesManager } from '../spaces_manager';
-import { SpacesGridPage } from './spaces_grid';
-import { ManageSpacePage } from './edit_space';
-import { Space } from '..';
+import type { SpacesManager } from '../spaces_manager';
 
 interface CreateParams {
   getStartServices: StartServicesAccessor<PluginsStart>;
@@ -35,108 +28,12 @@ export const spacesManagementApp = Object.freeze({
       order: 2,
       title,
 
-      async mount({ element, setBreadcrumbs, history }) {
-        const [
-          { notifications, i18n: i18nStart, application, chrome },
-          { features },
-        ] = await getStartServices();
-        const spacesBreadcrumbs = [
-          {
-            text: title,
-            href: `/`,
-          },
-        ];
+      async mount(params) {
+        const [coreStart, { features }] = await getStartServices();
 
-        chrome.docTitle.change(title);
+        const { renderApp } = await import('./render_app');
 
-        const SpacesGridPageWithBreadcrumbs = () => {
-          setBreadcrumbs(spacesBreadcrumbs);
-          return (
-            <SpacesGridPage
-              capabilities={application.capabilities}
-              getFeatures={features.getFeatures}
-              notifications={notifications}
-              spacesManager={spacesManager}
-              history={history}
-              getUrlForApp={application.getUrlForApp}
-            />
-          );
-        };
-
-        const CreateSpacePageWithBreadcrumbs = () => {
-          setBreadcrumbs([
-            ...spacesBreadcrumbs,
-            {
-              text: i18n.translate('xpack.spaces.management.createSpaceBreadcrumb', {
-                defaultMessage: 'Create',
-              }),
-            },
-          ]);
-
-          return (
-            <ManageSpacePage
-              capabilities={application.capabilities}
-              getFeatures={features.getFeatures}
-              notifications={notifications}
-              spacesManager={spacesManager}
-              history={history}
-              getUrlForApp={application.getUrlForApp}
-            />
-          );
-        };
-
-        const EditSpacePageWithBreadcrumbs = () => {
-          const { spaceId } = useParams<{ spaceId: string }>();
-
-          const onLoadSpace = (space: Space) => {
-            setBreadcrumbs([
-              ...spacesBreadcrumbs,
-              {
-                text: space.name,
-                href: `/edit/${encodeURIComponent(space.id)}`,
-              },
-            ]);
-          };
-
-          return (
-            <ManageSpacePage
-              capabilities={application.capabilities}
-              getFeatures={features.getFeatures}
-              notifications={notifications}
-              spacesManager={spacesManager}
-              spaceId={spaceId}
-              onLoadSpace={onLoadSpace}
-              history={history}
-              getUrlForApp={application.getUrlForApp}
-            />
-          );
-        };
-
-        render(
-          <i18nStart.Context>
-            <RedirectAppLinks application={application}>
-              <Router history={history}>
-                <Switch>
-                  <Route path={['', '/']} exact>
-                    <SpacesGridPageWithBreadcrumbs />
-                  </Route>
-                  <Route path="/create">
-                    <CreateSpacePageWithBreadcrumbs />
-                  </Route>
-                  <Route path="/edit/:spaceId">
-                    <EditSpacePageWithBreadcrumbs />
-                  </Route>
-                </Switch>
-              </Router>
-            </RedirectAppLinks>
-          </i18nStart.Context>,
-          element
-        );
-
-        return () => {
-          chrome.docTitle.reset();
-          unmountComponentAtNode(element);
-        };
+        return renderApp(coreStart, params, features, spacesManager);
       },
     } as RegisterManagementAppArgs;
   },
