+++ x-pack/plugins/spaces/public/nav_control/nav_control.tsx
@@ -5,11 +5,12 @@
  * 2.0.
  */
 
-import React from 'react';
+import React, { Suspense } from 'react';
 import ReactDOM from 'react-dom';
 import { CoreStart } from 'src/core/public';
-import { SpacesManager } from '../spaces_manager';
-import { NavControlPopover } from './nav_control_popover';
+import type { SpacesManager } from '../spaces_manager';
+
+const NavControlPopover = React.lazy(() => import('./nav_control_popover'));
 
 export function initSpacesNavControl(spacesManager: SpacesManager, core: CoreStart) {
   const I18nContext = core.i18n.Context;
@@ -22,13 +23,15 @@ export function initSpacesNavControl(spacesManager: SpacesManager, core: CoreSta
 
       ReactDOM.render(
         <I18nContext>
-          <NavControlPopover
-            spacesManager={spacesManager}
-            serverBasePath={core.http.basePath.serverBasePath}
-            anchorPosition="downLeft"
-            capabilities={core.application.capabilities}
-            navigateToApp={core.application.navigateToApp}
-          />
+          <Suspense fallback={<div />}>
+            <NavControlPopover
+              spacesManager={spacesManager}
+              serverBasePath={core.http.basePath.serverBasePath}
+              anchorPosition="downLeft"
+              capabilities={core.application.capabilities}
+              navigateToApp={core.application.navigateToApp}
+            />
+          </Suspense>
         </I18nContext>,
         targetDomElement
       );
