+++ x-pack/plugins/spaces/public/share_saved_objects_to_space/components/index.ts
@@ -5,6 +5,5 @@
  * 2.0.
  */
 
-export { ShareToSpaceFlyoutInternal } from './share_to_space_flyout_internal';
 export { getShareToSpaceFlyoutComponent } from './share_to_space_flyout';
 export { getLegacyUrlConflict } from './legacy_url_conflict';
