+++ x-pack/plugins/spaces/public/share_saved_objects_to_space/components/legacy_url_conflict.tsx
@@ -5,14 +5,20 @@
  * 2.0.
  */
 
-import React from 'react';
+import React, { Suspense } from 'react';
 import type { LegacyUrlConflictProps } from 'src/plugins/spaces_oss/public';
-import { LegacyUrlConflictInternal, InternalProps } from './legacy_url_conflict_internal';
+import type { InternalProps } from './legacy_url_conflict_internal';
+
+const LegacyUrlConflictInternal = React.lazy(() => import('./legacy_url_conflict_internal'));
 
 export const getLegacyUrlConflict = (
   internalProps: InternalProps
 ): React.FC<LegacyUrlConflictProps> => {
   return (props: LegacyUrlConflictProps) => {
-    return <LegacyUrlConflictInternal {...{ ...internalProps, ...props }} />;
+    return (
+      <Suspense fallback={<div />}>
+        <LegacyUrlConflictInternal {...{ ...internalProps, ...props }} />{' '}
+      </Suspense>
+    );
   };
 };
