+++ x-pack/plugins/spaces/public/share_saved_objects_to_space/index.ts
@@ -5,6 +5,5 @@
  * 2.0.
  */
 
-export { ShareSavedObjectsToSpaceService } from './share_saved_objects_to_space_service';
 export { getShareToSpaceFlyoutComponent, getLegacyUrlConflict } from './components';
 export { createRedirectLegacyUrl } from './utils';
