+++ x-pack/plugins/spaces/public/share_saved_objects_to_space/utils/redirect_legacy_url.ts
@@ -6,7 +6,7 @@
  */
 
 import { i18n } from '@kbn/i18n';
-import { firstValueFrom } from '@kbn/std';
+import { first } from 'rxjs/operators';
 import type { StartServicesAccessor } from 'src/core/public';
 import type { SpacesApiUi } from 'src/plugins/spaces_oss/public';
 import type { PluginsStart } from '../../plugin';
@@ -18,7 +18,7 @@ export function createRedirectLegacyUrl(
   return async function (path: string, objectNoun: string = DEFAULT_OBJECT_NOUN) {
     const [{ notifications, application }] = await getStartServices();
     const { currentAppId$, navigateToApp } = application;
-    const appId = await firstValueFrom(currentAppId$); // retrieve the most recent value from the BehaviorSubject
+    const appId = await currentAppId$.pipe(first()).toPromise(); // retrieve the most recent value from the BehaviorSubject
 
     const title = i18n.translate('xpack.spaces.shareToSpace.redirectLegacyUrlToast.title', {
       defaultMessage: `We redirected you to a new URL`,
