+++ x-pack/plugins/spaces/public/space_list/space_list_internal.tsx
@@ -142,3 +142,5 @@ export const SpaceListInternal = ({
     </EuiFlexGroup>
   );
 };
+
+export default SpaceListInternal;
