+++ x-pack/plugins/spaces/public/space_selector/space_selector_app.tsx
@@ -7,7 +7,7 @@
 
 import { StartServicesAccessor, ApplicationSetup, AppMountParameters } from 'src/core/public';
 import { i18n } from '@kbn/i18n';
-import { SpacesManager } from '../spaces_manager';
+import type { SpacesManager } from '../spaces_manager';
 
 interface CreateDeps {
   application: ApplicationSetup;
